+++ CHANGELOG.md
@@ -64,6 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Arc3D: line_intersections, linesegment_intersections
 * new module utils: intersections -> circle_3d_linesegment_intersections
 * hash for Frame2D
+* Ellipse3D: point_belongs, abscissa, length, to_2d
 * CylindricalSurface3D: point_on_surface, is_coincident, arcellipse3d_to_2d
 
 ### Fixed
@@ -86,6 +87,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Plane3D: plane_intersections, is_coindident
 * contour2d: ordering_primitives, order_primitives
 * infinite primitive offset of linesegment
+* Ellispe3D: discretization_points
 
 ### Performance improvements
 
@@ -97,6 +99,7 @@ time we have to verify if a point is inside
 ### Refactorings
 * LineSegment3D: intersections
 * Line2D: sort_points_along_line
+* Line3D: intersections
 
 
 ### Unittests
@@ -140,6 +143,7 @@ time we have to verify if a point is inside
 * Circle3D: discretization_points
 * Arc3D: line_intersections, linesegment_intersections
 * Contour2D: ordering_contour, is_ordered, order_contour
+* Ellipse3D: point_belongs, abscissa, length, to_2d, discretization_points
 * CylindricalSurface3D: point_on_surface, is_coincident
 
 ## v0.6.1 [12/13/2022]
@@ -193,7 +197,6 @@ time we have to verify if a point is inside
 * Surface2D: cut_by_line
 * Circle3D: to_step
 * ArcEllipse3D.to_2d()
-
 * infinite primitive offset of linesegment
 
 ### Performance improvements